requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
cffi==2.0.0
click==8.3.1
fastapi==0.129.0
h11==0.16.0
idna==3.11
joblib==1.4.2
pandas==2.2.0
pycparser==3.0
scikit-learn==1.5.0
pydantic==2.12.5
pydantic_core==2.41.5
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.41.0
